Move Meta to Facades namespace
