+++ packages/framework/src/Framework/Features/Metadata/GlobalMetadataBag.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Features\Metadata;
 
+use Hyde\Facades\Meta;
 use Hyde\Framework\Concerns\HydePage;
 use Hyde\Framework\Services\RssFeedService;
 use Hyde\Helpers\Features;
-use Hyde\Helpers\Meta;
 use Hyde\Hyde;
 use Illuminate\Support\Facades\View;
