+++ packages/framework/src/Framework/Features/Metadata/PageMetadataBag.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Features\Metadata;
 
+use Hyde\Facades\Meta;
 use Hyde\Framework\Concerns\HydePage;
-use Hyde\Helpers\Meta;
 use Hyde\Pages\MarkdownPost;
 
 class PageMetadataBag extends MetadataBag
